refactor: split WearGameRepository into focused repositories
