+++ android-app/wear/src/androidTest/java/dev/convocados/wear/e2e/ScoreUpdateE2ETest.kt
@@ -15,6 +15,8 @@ import dev.convocados.wear.data.local.dao.PendingScoreDao
 import dev.convocados.wear.data.local.dao.WearGameDao
 import dev.convocados.wear.data.local.dao.WearHistoryDao
 import dev.convocados.wear.data.repository.WearGameRepository
+import dev.convocados.wear.data.repository.WearScoreRepository
+import dev.convocados.wear.data.repository.WearScoreRepository
 import io.ktor.client.*
 import io.ktor.client.call.*
 import io.ktor.client.engine.okhttp.*
@@ -80,6 +82,8 @@ class ScoreUpdateE2ETest {
     private lateinit var historyDao: WearHistoryDao
     private lateinit var pendingScoreDao: PendingScoreDao
     private lateinit var repository: WearGameRepository
+    private lateinit var scoreRepository: WearScoreRepository
+    private lateinit var scoreRepository: WearScoreRepository
 
     private val json = Json {
         ignoreUnknownKeys = true
@@ -113,7 +117,8 @@ class ScoreUpdateE2ETest {
         historyDao = db.historyDao()
         pendingScoreDao = db.pendingScoreDao()
 
-        repository = WearGameRepository(apiClient, gameDao, historyDao, pendingScoreDao)
+        repository = WearGameRepository(apiClient, gameDao, historyDao)
+        scoreRepository = WearScoreRepository(apiClient, historyDao, pendingScoreDao)
     }
 
     @After
@@ -230,7 +235,7 @@ class ScoreUpdateE2ETest {
         val newScoreOne = (historyEntry.scoreOne ?: 0) + 1
         val newScoreTwo = historyEntry.scoreTwo ?: 0
 
-        val submitResult = repository.submitScore(
+        val submitResult = scoreRepository.submitScore(
             eventId = targetGame.id,
             historyId = historyEntry.id,
             scoreOne = newScoreOne,
