+++ android-app/wear/src/test/java/dev/convocados/wear/data/repository/WearGameRepositoryTest.kt
@@ -6,14 +6,9 @@ import dev.convocados.wear.data.api.GameHistory
 import dev.convocados.wear.data.api.MyGamesResponse
 import dev.convocados.wear.data.api.PaginatedHistory
 import dev.convocados.wear.data.api.WearApiClient
-import dev.convocados.wear.data.local.dao.PendingRosterChangeDao
-import dev.convocados.wear.data.local.dao.PendingScoreDao
 import dev.convocados.wear.data.local.dao.WearGameDao
 import dev.convocados.wear.data.local.dao.WearHistoryDao
-import dev.convocados.wear.data.local.dao.WearPlayerDao
-import dev.convocados.wear.data.local.entity.PendingScoreEntity
 import dev.convocados.wear.data.local.entity.WearGameEntity
-import dev.convocados.wear.data.local.entity.WearHistoryEntity
 import io.mockk.*
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.runTest
@@ -26,18 +21,13 @@ class WearGameRepositoryTest {
     private val client = mockk<WearApiClient>()
     private val gameDao = mockk<WearGameDao>(relaxed = true)
     private val historyDao = mockk<WearHistoryDao>(relaxed = true)
-    private val pendingScoreDao = mockk<PendingScoreDao>(relaxed = true)
-    private val playerDao = mockk<WearPlayerDao>(relaxed = true)
-    private val pendingRosterChangeDao = mockk<PendingRosterChangeDao>(relaxed = true)
 
     private lateinit var repository: WearGameRepository
 
     @Before
     fun setup() {
-        repository = WearGameRepository(client, gameDao, historyDao, pendingScoreDao, playerDao, pendingRosterChangeDao)
+        repository = WearGameRepository(client, gameDao, historyDao)
     }
-
-    // ── observeGames ─────────────────────────────────────────────────────
 
     @Test
     fun `observeGames delegates to gameDao`() = runTest {
@@ -49,8 +39,6 @@ class WearGameRepositoryTest {
             awaitComplete()
         }
     }
-
-    // ── refreshGames ─────────────────────────────────────────────────────
 
     @Test
     fun `refreshGames fetches from API and updates dao`() = runTest {
@@ -77,8 +65,6 @@ class WearGameRepositoryTest {
         assertEquals("Network error", result.exceptionOrNull()?.message)
     }
 
-    // ── refreshHistory ───────────────────────────────────────────────────
-
     @Test
     fun `refreshHistory fetches and caches history`() = runTest {
         val history = PaginatedHistory(
@@ -103,78 +89,6 @@ class WearGameRepositoryTest {
         assertTrue(result.isFailure)
     }
 
-    // ── submitScore ──────────────────────────────────────────────────────
-
-    @Test
-    fun `submitScore updates local cache optimistically`() = runTest {
-        coEvery { client.patch<GameHistory>(any(), any()) } returns GameHistory(
-            "h1", "2025-01-01T10:00:00Z", "played", 5, 3, "Red", "Blue", true,
-        )
-
-        val result = repository.submitScore("e1", "h1", 5, 3, "Red", "Blue")
-
-        assertTrue(result.isSuccess)
-        coVerify { historyDao.updateScore("h1", 5, 3) }
-    }
-
-    @Test
-    fun `submitScore queues pending score on network failure`() = runTest {
-        coEvery { client.patch<GameHistory>(any(), any()) } throws Exception("Offline")
-
-        val result = repository.submitScore("e1", "h1", 5, 3, "Red", "Blue")
-
-        // Returns failure to signal offline, but score is queued for later sync
-        assertTrue(result.isFailure)
-        coVerify { pendingScoreDao.insert(match { it.eventId == "e1" && it.scoreOne == 5 }) }
-    }
-
-    // ── syncPendingScores ────────────────────────────────────────────────
-
-    @Test
-    fun `syncPendingScores syncs all pending and deletes them`() = runTest {
-        val pending = listOf(
-            PendingScoreEntity(1, "e1", "h1", 3, 2, "A", "B"),
-            PendingScoreEntity(2, "e2", "h2", 1, 0, "C", "D"),
-        )
-        coEvery { pendingScoreDao.getAll() } returns pending
-        coEvery { client.patch<GameHistory>(any(), any()) } returns GameHistory(
-            "h1", "2025-01-01T10:00:00Z", "played", 3, 2, "A", "B", true,
-        )
-
-        val synced = repository.syncPendingScores()
-
-        assertEquals(2, synced)
-        coVerify(exactly = 2) { pendingScoreDao.delete(any()) }
-        coVerify { pendingScoreDao.deleteStale() }
-    }
-
-    @Test
-    fun `syncPendingScores increments retry on failure`() = runTest {
-        val pending = listOf(PendingScoreEntity(1, "e1", "h1", 3, 2, "A", "B"))
-        coEvery { pendingScoreDao.getAll() } returns pending
-        coEvery { client.patch<GameHistory>(any(), any()) } throws Exception("Still offline")
-
-        val synced = repository.syncPendingScores()
-
-        assertEquals(0, synced)
-        coVerify { pendingScoreDao.incrementRetry(1) }
-        coVerify { pendingScoreDao.deleteStale() }
-    }
-
-    // ── observePendingCount ──────────────────────────────────────────────
-
-    @Test
-    fun `observePendingCount delegates to pendingScoreDao`() = runTest {
-        coEvery { pendingScoreDao.observeCount() } returns flowOf(3)
-
-        repository.observePendingCount().test {
-            assertEquals(3, awaitItem())
-            awaitComplete()
-        }
-    }
-
-    // ── getGame ──────────────────────────────────────────────────────────
-
     @Test
     fun `getGame returns cached game`() = runTest {
         val game = makeGame("e1")
@@ -189,8 +103,6 @@ class WearGameRepositoryTest {
 
         assertNull(repository.getGame("missing"))
     }
-
-    // ── Helpers ──────────────────────────────────────────────────────────
 
     private fun makeGame(id: String) = WearGameEntity(
         id = id,
